requirements.txt
pylint
mysql-connector-python
